requirements/base.txt
# Base Runtime Dependencies
# =========================
# Core dependencies needed for inference/scoring

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data Manipulation
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data Sources
requests>=2.31.0
yfinance>=0.2.28

# PIT-Safe Storage & Fast Replay
pyarrow>=14.0.0
duckdb>=0.9.0

# Trading Calendar (correct cutoffs/trading days)
exchange-calendars>=4.2.0

# Data Validation
pydantic>=2.0.0

# Logging
loguru>=0.7.0

# Utilities
tqdm>=4.65.0
pytz>=2023.3
python-dateutil>=2.8.2

# Caching
joblib>=1.3.0

